BigInteger on object_id
Adding BigInteger as the object_id to support PK's coming from facebook

+++ taggit/models.py
@@ -117,7 +117,7 @@ class TaggedItemBase(ItemBase):
 
 
 class GenericTaggedItemBase(ItemBase):
-    object_id = models.IntegerField(verbose_name=_('Object id'), db_index=True)
+    object_id = models.BigIntegerField(verbose_name=_('Object id'), db_index=True)
     if django.VERSION < (1, 2):
         content_type = models.ForeignKey(
             ContentType,
